refactor: add function to fetch rewards address in a hook
