+++ packages/react-app-revamp/components/_pages/Create/pages/ContestRewards/index.tsx
@@ -11,12 +11,10 @@ const CreateContestRewards = () => {
   const { reset: clearContestData, isSuccess: contestDeployed } = useDeployContestStore(state => state);
   const {
     displayCreatePool,
-    isLoading: isRewardsModuleDeploying,
     cancel: cancelCreateRewardsPool,
     reset: clearRewardsData,
   } = useDeployRewardsStore(state => state);
   const { setShowRewards } = useShowRewardsStore(state => state);
-  const { isLoading: isFundingRewardsDeploying } = useFundRewardsStore(state => state);
   const { cancel: cancelFundingPool } = useFundRewardsStore(state => state);
 
   const [isOpen, setIsOpen] = useState(true);
