+++ packages/react-app-revamp/components/_pages/Rewards/components/Fund/index.tsx
@@ -1,15 +1,12 @@
 import MultiStepToast, { ToastMessage } from "@components/UI/MultiStepToast";
 import { toastError } from "@components/UI/Toast";
-import { chains } from "@config/wagmi";
-import getContestContractVersion from "@helpers/getContestContractVersion";
 import useFundRewardsModule from "@hooks/useFundRewards";
 import { useFundRewardsStore } from "@hooks/useFundRewards/store";
-import { fetchToken, readContract } from "@wagmi/core";
+import useRewardsModule from "@hooks/useRewards";
+import { fetchToken } from "@wagmi/core";
 import { ethers } from "ethers";
-import { useRouter } from "next/router";
 import { FC, useRef } from "react";
 import { toast } from "react-toastify";
-import { Abi } from "viem";
 import { useAccount } from "wagmi";
 import CreateRewardsFundingPoolSubmit from "./components/Buttons/Submit";
 import CreateRewardsFundPool from "./components/FundPool";
@@ -19,13 +16,8 @@ interface CreateRewardsFundingProps {
 }
 
 const CreateRewardsFunding: FC<CreateRewardsFundingProps> = ({ isFundingForTheFirstTime = true }) => {
-  const { asPath } = useRouter();
-  const contestAddress = asPath.split("/")[3];
-  const contestChainName = asPath.split("/")[2];
-  const chainId = chains.filter(
-    chain => chain.name.toLowerCase().replace(" ", "") === contestChainName.toLowerCase(),
-  )?.[0]?.id;
   const { sendFundsToRewardsModuleV3 } = useFundRewardsModule();
+  const { getContestRewardsAddress } = useRewardsModule();
   const { address } = useAccount();
   const { rewards, setCancel } = useFundRewardsStore(state => state);
   const toastIdRef = useRef<string | number | null>(null);
@@ -36,9 +28,9 @@ const CreateRewardsFunding: FC<CreateRewardsFundingProps> = ({ isFundingForTheFi
     const populatedRewardsPromises = rewards.map(async reward => {
       if (reward.amount === "") return null;
 
-      const rewardsContractAddress = await getRewardsModuleAddress();
+      const rewardsContractAddress = await getContestRewardsAddress();
 
-      if (rewardsContractAddress === "0x0000000000000000000000000000000000000000") {
+      if (!rewardsContractAddress || rewardsContractAddress === "0x0000000000000000000000000000000000000000") {
         toastError("there is no rewards module for this contest!");
         return;
       }
@@ -91,29 +83,6 @@ const CreateRewardsFunding: FC<CreateRewardsFundingProps> = ({ isFundingForTheFi
         icon: false,
       },
     );
-  };
-
-  const getRewardsModuleAddress = async () => {
-    const { abi: abiContest } = await getContestContractVersion(contestAddress, chainId);
-
-    if (abiContest === null) {
-      toastError(`This contract doesn't exist on ${contestChainName}.`);
-      return;
-    }
-
-    try {
-      const contestRewardModuleAddress = (await readContract({
-        address: contestAddress as `0x${string}`,
-        abi: abiContest as unknown as Abi,
-        chainId,
-        functionName: "officialRewardsModule",
-      })) as string;
-
-      return contestRewardModuleAddress;
-    } catch (error) {
-      toastError("failed to fetch rewards module address");
-      return;
-    }
   };
 
   const onCancelFundingPool = () => {
