+++ packages/react-app-revamp/hooks/useDeployRewards/store.tsx
@@ -1,5 +1,4 @@
-import { createContext, useContext, useRef } from "react";
-import { createStore, useStore } from "zustand";
+import { create } from "zustand";
 
 interface DeployRewardsState {
   ranks: number[];
@@ -31,58 +30,38 @@ interface DeployRewardsState {
   reset: () => void;
 }
 
-export const createDeployRewardsStore = () =>
-  createStore<DeployRewardsState>(set => {
-    const initialState = {
-      ranks: [],
-      shares: [],
-      isLoading: false,
-      isSuccess: false,
-      isError: false,
-      cancel: false,
-      validationError: {
-        uniqueRanks: undefined,
-        zeroProportion: undefined,
-        invalidTotal: undefined,
-        duplicateRank: undefined,
-      },
-      displayCreatePool: true,
-      deployRewardsData: {
-        hash: "",
-        address: "",
-      },
-    };
+export const useDeployRewardsStore = create<DeployRewardsState>(set => {
+  const initialState = {
+    ranks: [],
+    shares: [],
+    isLoading: false,
+    isSuccess: false,
+    isError: false,
+    cancel: false,
+    validationError: {
+      uniqueRanks: undefined,
+      zeroProportion: undefined,
+      invalidTotal: undefined,
+      duplicateRank: undefined,
+    },
+    displayCreatePool: true,
+    deployRewardsData: {
+      hash: "",
+      address: "",
+    },
+  };
 
-    return {
-      ...initialState,
-      setValidationError: error => set({ validationError: error }),
-      setIsLoading: isLoading => set({ isLoading }),
-      setIsSuccess: isSuccess => set({ isSuccess }),
-      setIsError: isError => set({ isError }),
-      setCancel: cancel => set({ cancel: cancel }),
-      setDeployRewardsData: (hash, address) => set(state => ({ deployRewardsData: { hash, address } })),
-      setRanks: ranks => set({ ranks }),
-      setShares: shares => set({ shares }),
-      setDisplayCreatePool: displayCreatePool => set({ displayCreatePool }),
-      reset: () => set({ ...initialState }),
-    };
-  });
-
-export const DeployRewardsContext = createContext<ReturnType<typeof createDeployRewardsStore> | null>(null);
-
-export function DeployRewardsWrapper({ children }: { children: React.ReactNode }) {
-  const storeRef = useRef<ReturnType<typeof createDeployRewardsStore>>();
-  if (!storeRef.current) {
-    storeRef.current = createDeployRewardsStore();
-  }
-  return <DeployRewardsContext.Provider value={storeRef.current}>{children}</DeployRewardsContext.Provider>;
-}
-
-export function useDeployRewardsStore<T>(selector: (state: DeployRewardsState) => T) {
-  const store = useContext(DeployRewardsContext);
-  if (store === null) {
-    throw new Error("Missing DeployRewards in the tree");
-  }
-  const value = useStore(store, selector);
-  return value;
-}
+  return {
+    ...initialState,
+    setValidationError: error => set({ validationError: error }),
+    setIsLoading: isLoading => set({ isLoading }),
+    setIsSuccess: isSuccess => set({ isSuccess }),
+    setIsError: isError => set({ isError }),
+    setCancel: cancel => set({ cancel: cancel }),
+    setDeployRewardsData: (hash, address) => set(state => ({ deployRewardsData: { hash, address } })),
+    setRanks: ranks => set({ ranks }),
+    setShares: shares => set({ shares }),
+    setDisplayCreatePool: displayCreatePool => set({ displayCreatePool }),
+    reset: () => set({ ...initialState }),
+  };
+});
