+++ packages/react-app-revamp/pages/contest/new/index.tsx
@@ -1,7 +1,5 @@
 import CreateFlow from "@components/_pages/Create";
 import { ContractFactoryWrapper } from "@hooks/useContractFactory";
-import { DeployRewardsWrapper } from "@hooks/useDeployRewards/store";
-import { FundRewardsWrapper } from "@hooks/useFundRewards/store";
 import { RewardsWrapper } from "@hooks/useRewards/store";
 import type { NextPage } from "next";
 import Head from "next/head";
@@ -15,13 +13,9 @@ const Page: NextPage = () => {
       </Head>
 
       <ContractFactoryWrapper>
-        <DeployRewardsWrapper>
-          <RewardsWrapper>
-            <FundRewardsWrapper>
-              <CreateFlow />
-            </FundRewardsWrapper>
-          </RewardsWrapper>
-        </DeployRewardsWrapper>
+        <RewardsWrapper>
+          <CreateFlow />
+        </RewardsWrapper>
       </ContractFactoryWrapper>
     </>
   );
